Implement TransformedKNeighborsMixin
Estimators that use a transformer need to run transform when kneighbors
is called with new data. The TransformedKNeighborsMixin overrides
kneighbors to run that transform. This mixin may be able to handle
running the transform when fitting and predicting with those
estimators, but that would require handling custom parameters like
CCA parameters for GNN, and may make function signatures less
readable. Worth considering further at a later date.