Add package path to hatch config
See pypa/hatch#845. This was necessary because our project name
is different from our package name, so hatch wasn't able to
locate it in the default path. Note that if the shell has been
built before making this change, it will need to be manually
rebuilt by running `hatch env prune`.