Subclass estimators from KNeighborsRegressor #13 (#14)
To allow fitting and predicting with continuous multi-output data,
estimators need to subclass from KNeighborsRegressor instead of
KNeighborsClassifier. This makes that change and renames the
IDNeighborsClassifier accordingly. One test is added to confirm
the basic functionality of fit and predict with multi-output data,
but more comprehensive tests will likely be needed in the future.

Additionally, the kneighbor_ids method was removed from the
IDNeighborsRegressor because, following discussion in #2, it is
no longer needed. The class was left as a placeholder pending
further discussion about handling plot IDs via kneighbors.