Simplify fit parameters for `CCATransformer` and `CCorATransformer` #30 (#31)
Transformers no longer take `spp` parameters when fitting. Instead,
estimators now choose the appropriate `y_fit` data to pass to the
transformer. This simplifies the responsibility of transformers and
avoids some error checking that would have been needed otherwise.

The `spp` parameter to the estimators was renamed to `y_fit` to be more
general.