Test all estimators' predict method for Moscow test data #26 (#32)
We exported data from yaImpute for the five core estimators at k=5
for both unweighted (equal-weighting to all neighbors) and weighted
(neighbors are weighted by their inverse distance) predictions of the
y attributes.  The tests assert that the predict method in each
estimator is returning the same values.  Note that the constructor
of each estimator uses a custom function (yaimpute_weights) to the
weights parameter in order to match the weighting scheme of yaImpute.

This commit fixes #26.