Set up transformers for dataframe compatibility #20
All transformers now support `get_feature_names_out` and `set_output`
methods. The first method was manually implemented for CCA and CCorA
and was inherited from `OneToOneFeatureMixin` for Mahalanobis. The
second method was automatically available once `get_feature_names_out`
was implemented, because all transformers subclass `BaseEstimator` and
indirectly `_SetOutputMixin`. To get `get_feature_names_out` working,
this also implements `_n_features_out` properties for CCA and CCorA.

Tests for these new features are included to ensure that the length of
feature names matches the output features of each transformer, and that
set_output is callable.

Tests are passing, but warnings are raised when estimators are fit with
dataframes. This will be fixed once we use `set_output` to set the
transformer mode in our estimators and remove the `NamedFeatureArray`
hack.