Finalize dataframe compatibility #20
`NamedFeatureArray`, which was used to trick sklearn into storing
feature names from array inputs, is now removed. Instead, we use the
`set_output` method on transformers to ensure that they pass dataframes
through to allow estimators to store feature names.

`feature_names_in_` was overridden for all transformed estimators to
return feature names from the transformer rather than the estimator.
This means that the property will return the names of features that
the user passed in to fit the estimator, rather than the transformed
features that were used internally (e.g. cca0, cca1, etc).

Overriding `feature_names_in_` caused `_check_feature_names` to fail
when called during fitting because the `feature_names_in_` intentionally
mismatch the names seen during fit. To overcome that, we override that
method to remove the affected check. We still need to handle warnings if
feature names are incorrectly missing, so we currently borrow part of the
implementation for that method from sklearn (BSD-3 license).

This commit modifies some calls to `_validate_data` from the previous
commit to avoid overriding X. This is done because `_validate_data` casts
to array, which can cause issues when a transformer calls a subsequent
transformer (e.g. MahalanobisTransformer calls StandardScalerWithDOF)
with a dataframe input, as feature names will not match between the
transformers, leading to user warnings when predicting. Instead, X is
explicitly cast when needed using `np.asarray` and validation is
always done without overriding X. Note that `_validate_data` must be
called by all transformers that do not call it via a superclass because
this indirectly stores feature names.

While implementing this, the difficulty of tracking what output types
are expected from what transformers and estimators with what inputs and
configuration options became VERY clear, so we now have some basic
"consistency" tests that compare all of our transformers and estimators
with a comparable sklearn implementation to check output types and attrs
under a range of situations.

A very small change is that parametrized estimator checks are now passed
classes instead of instances because this leads to more helpful pytest
errors.