Dataframe and `set_output` compatibility for transformers (#34)
All transformers now support `get_feature_names_out` and `set_output`
methods. The first method was manually implemented for CCA and CCorA
and was inherited from `OneToOneFeatureMixin` for Mahalanobis. The
second method was automatically available once `get_feature_names_out`
was implemented, because all transformers subclass `BaseEstimator` and
indirectly `_SetOutputMixin`. To get `get_feature_names_out` working,
this also implements `_n_features_out` properties for CCA and CCorA.

`NamedFeatureArray`, which was used to trick sklearn into storing
feature names from array inputs, is now removed. Instead, we use the
`set_output` method on transformers to ensure that they pass dataframes
through to allow estimators to store feature names.

`feature_names_in_` was overridden for all transformed estimators to
return feature names from the transformer rather than the estimator.
This means that the property will return the names of features that
the user passed in to fit the estimator, rather than the transformed
features that were used internally (e.g. cca0, cca1, etc).

Overriding `feature_names_in_` caused `_check_feature_names` to fail
when called during fitting because the `feature_names_in_` intentionally
mismatch the names seen during fit. To overcome that, we override that
method to remove the affected check. We still need to handle warnings if
feature names are incorrectly missing, so we currently borrow part of the
implementation for that method from sklearn (BSD-3 license).

Because we are now compatible with and rely on the `set_output` API,
scikit-learn is pinned to >=1.2, which is the release that introduced that
feature.

This also adds a number of tests to check consistency of attributes and
output data types between our transformers and estimators and reference
transformers and estimators from sklearn.