Allow for user to specify reduced number of components (axes) #33 (#40)
This feature introduces a new mixin (`ComponentReducerMixin`) to handle dimension reduction for transformers which create "axes" as linear combinations of their X features (presently `CCATransformer` and `CCorATransformer`).  For these transformers, there is an upper limit on the number of axes that can be used, either by matrix rank or using tests to determine significance of axes.  The user can specify a hyperparameter (`n_components`) on these transformers
to use fewer than the maximum number of axes, which in turn changes the n-dimensional space for NN finding.

A few relevant changes:
- `CCA` and `CCorA` ordination classes now have a `max_components` property to ensure consistency so that either can be used in conjunction with `ComponentReducerMixin`
- `CCA` has been refactored to combine the coefficients and axis_weights into a single matrix, which has been moved into the `projector` method (similar to `CCorA`).
- Both `CCA` and `CCorA`'s `projector` method now accommodates a parameter for `n_components`
- Tests have been added to ensure that transformed X data has the correct number of components and raise errors if the user specified value is outside the range 0 - max_components.
- Tests have been added to ensure that `GNNRegressor` and `MSNRegressor` use the correct number of components in `kneighbors` and `predict`.


---------

Co-authored-by: Aaron Zuspan <50475791+aazuspan@users.noreply.github.com>