Fix side-effects caused by global `set_config`
Some of our tests used `set_config` to set the global output mode,
which caused side effects to other tests. Somehow this never caused
issues with our tests, but it broke several of the estimator checks
if our checks ran first.

To fix that, we now use `config_context` to ensure that any changes
to global config are temporary. Note that in those tests, we use
`config_context` with an empty config if we don't want to make any
changes - that was the cleanest solution I could come up as there's
no good way I know of to conditionally run a context manager.