Fix MSNRegressor()-check_n_features_in_
This check was failing because the `n_features_in_` attribute does
not match the number of features that were used to fit the model.
This is essentially the same problem we faced with
`feature_names_in_`, where the model stores metadata about the
features after transformation, when it should instead be storing
data on the original, untransformed features.

To fix that, we use the same solution we did with `feature_names_in_`,
which is to 1) add a `n_features_in_` property that returns the
number of untransformed features and 2) override `_check_n_features`
to prevent errors. As with `feature_names_in_`, any real errors that
would be caught by the check are instead caught by the transformer
during `_apply_transform`.