Fix estimator checks (#44)
This gets all of our transformers and estimators passing the sklearn
estimator checks, or xfailing in the case of incompatible checks. The 
main change to accomplish that was adding more validation checks
and populating relevant tags.

Other notable changes are below:

- 1D y arrays caused StandardScalerWithDOF.fit_transform to fail
due to an inherited check with `ensure_2d=True`. Now, 1D y arrays
are reshaped to 2D before scaling, following the pattern of
`sklearn.cross_decomposition.CCA`.
- The `MSNRegressor.n_features_in_` attribute does
not match the number of features that were used to fit the model.
This is essentially the same problem we faced with
`feature_names_in_`, where the model stores metadata about the
features after transformation, when it should instead be storing
data on the original, untransformed features. To fix that, we use the 
same solution we did with `feature_names_in_`, which is to 1) add an 
`n_features_in_` property that returns the number of untransformed 
features and 2) override `_check_n_features` to prevent errors. As 
with `feature_names_in_`, any real errors that would be caught by the 
check are instead caught by the transformer during `_apply_transform`.
* StandardScalerWithDOF is not invariant to zero sample weights, so for
simplicity we're just removing the argument as it's unused by any
of our estimators.

Co-authored by: Matt Gregory <matt.gregory@oregonstate.edu>