Add IndependentPredictionMixin and use as base for all estimators #45
This mixin provides two methods (`predict_independent` and
`score_independent`) that don't include the query point itself when
finding nearest neighbors.  These methods are similar in functionality
to calling `KNeighborsRegressor.kneighbors()` without passing `X`.  In
this way, only "independent" neighbors are used in calculating
predictions and scores.

This PR is intended to fix #45.