Migrate estimator_check xfails from tags
With sklearn 1.6, xfails for estimator checks are now defined as args
to the test functions instead of as estimator tags. Maintaining backwards
compatibility in the test suite would require a lot of added complexity,
so I'm proposing that we just pin the test environment >= 1.6. I also
added a few new failing estimator checks to the xfails.

Note that this is just the first step towards 1.6 compatibility, and there
are lots of failing tests still.