Make estimators store transformed feature names and numbers
Previously, we made estimators return the raw features used to fit
their underlying transformers, rather than the transformed features
that were actually used to fit the estimator. I think this was a
reasonable goal to avoid exposing that internal detail to users,
but it broke some basic assumptions of sklearn and required hacky
workarounds to avoid failing checks run by superclasses during
fitting and prediction (i.e. overriding the checks with no-ops).

The switch in sklearn 1.6 from method checks to function checks
broke those workarounds, and would require more complex workarounds
to keep the same functionality. I'm suggesting that we give in and
just follow the sklearn expectation where the estimator stores the
transformed features that were used to fit it, and the transformer
stores the raw features that were used to fit it. This does reveal
the implementation to the user if they try to access those
attributes, but I think we were fighting a losing battle trying to
make those attributes serve two purposes.

In the future, we could expose new attributes to make the raw features
directly accessible (e.g. `raw_features_in_`), or just point users
to `estimator.transformer_.n_features_in_` in documentation.