Scikit-learn 1.6 compatibility (#79)
* Migrate estimator_check xfails from tags

With sklearn 1.6, xfails for estimator checks are now defined as args
to the test functions instead of as estimator tags. Maintaining backwards
compatibility in the test suite would require a lot of added complexity,
so I'm proposing that we just pin the test environment >= 1.6. I also
added a few new failing estimator checks to the xfails.

Note that this is just the first step towards 1.6 compatibility, and there
are lots of failing tests still.

* Migrate from _more_tags to __sklearn_tags__

Estimator tags are stored differently in sklearn 1.6. This is an
attempt to become compliant with the new spec and try to 1) limit
redundancy between specified tags and the inherited defaults and
2) more fully define the tags, but I wouldn't guarantee these are
exhaustive.

* Wrap `validate_data`

sklearn 1.6 moved BaseEstimator._validate_data to a public function.
For backwards compatibility, this adds a wrapper around that which
calls the function or estimator method depending on availability.

* Remove placeholder license to fix hatch build

Not sure the underlying cause (hatch update?), but the placeholder
license string was causing hatch builds to fail when they attempted
to validate the license.

* Pin sklearn > 1.6 for testing compatibility

sklearn 1.6 changed estimator check xfails from an estimator tag
to a parameter in the testing function. Maintaining backwards
compatibility for testing would have been complex, so unless
there's a strong need to test with older versions I suggest we
test with >= 1.6 for now.

* Typo

* Docstring clarification

* Make estimators store transformed feature names and numbers

Previously, we made estimators return the raw features used to fit
their underlying transformers, rather than the transformed features
that were actually used to fit the estimator. I think this was a
reasonable goal to avoid exposing that internal detail to users,
but it broke some basic assumptions of sklearn and required hacky
workarounds to avoid failing checks run by superclasses during
fitting and prediction (i.e. overriding the checks with no-ops).

The switch in sklearn 1.6 from method checks to function checks
broke those workarounds, and would require more complex workarounds
to keep the same functionality. I'm suggesting that we give in and
just follow the sklearn expectation where the estimator stores the
transformed features that were used to fit it, and the transformer
stores the raw features that were used to fit it. This does reveal
the implementation to the user if they try to access those
attributes, but I think we were fighting a losing battle trying to
make those attributes serve two purposes.

In the future, we could expose new attributes to make the raw features
directly accessible (e.g. `raw_features_in_`), or just point users
to `estimator.transformer_.n_features_in_` in documentation.

* Handle renamed `ensure_all_finite`

The compatibility wrapper needs to use `force_all_finite` instead of
`ensure_all_finite` for backwards compatibility with scikit-learn < 1.6.0.
I also made the keyword arguments more strict, since this is just for our
internal use.

* Remove redundant xfail checks

* Add future reminder

* Remove new tags for now

This will probably need to be re-added, but the goal is to keep
this PR focused on updating the format of the tags rather than the
content. We probably need to give some more thought to how tags
should be inherited to reduce duplication between estimators as well.