Finalize estimator tags
- CCA now correctly requires positive targets
- CCorA now correctly requires targets
- All estimators now explicitly prohibit sparse data. This is a
temporary measure to fix failing checks. A closer investigation of
sparse support is needed.
- The addition of `target_tags.required` to CCorA triggered an
estimator check `check_requires_y_none` which looks for an
expected error message when passed `y=None`. That error is now
correctly raised by passing y into `_validate_data`. This removed
the need to separately check the y array. Note that we don't mutate
`X` with the validation check as this will strip feature names that
must be passed to `StandardScalerWithDOF`.