Update test data files to match scikit-learn kneighbors
This commit is necessary to account for small changes in ordering of
neighbor IDs in the case of tied distances when running through
KNeighborsRegressor.kneighbors.  There were a few records where IDs
swapped locations when distances were identical.  This was causing tests
to fail when compared against original output from yaImpute.

We replace the neighbor files for both reference and target output after
verifying that distances were identical for these records and k
position.  Correspondingly, the predicted output for these records
needed to change.  It's important to note that neighbor distance files
did not need to be updated.

With these updated files, tests are passing locally for all tests in
test_port.py.