WIP: Full parameter sets for two base estimators
This is a interim commit to provide the full set of
`RandomForestRegressor` and `KNeighborsRegressor` parameters to the
`RFNNRegressor`.  Ideally, we would like to pass dictionaries with these
two sets of options, but are running into issues when sticking to the
model that an estimator's `__init__` should only set, and not modify,
its parameters.  This commit has a lot of boilerplate, but more closely
follows patterns in `scikit-learn`.