Modify get_feature_dtypes to return promoted dtypes when necessary
In an earlier commit, get_feature_dtypes returned a mix of dtypes based
on whether the object passed was a pandas object or a numpy object.
Now, we consistently return numpy dtypes that are the smallest type that
can hold all elements in a sequence.  The exception to this rule is when
a sequence is a pd.CategoricalDtype field, in which case we retain this
information.