Add internal _set_targets method to handle target standardization.
Because the targets (y) may be a mix of dtypes by column, standard
sklearn validation doesn't work as expected.  We chose to validation
features (X) separately in `fit`, and use the `_set_targets` function
to handle each target in y, first validating that there are no NaN-like
data as well as ensuring no mixed data of numeric and non-numeric data.
Finally, targets are captured as a list of 1-d numpy arrays with dtype
set to the smallest type capable of storing all data elements.