Auto-assign RF type based on feature data type (#91)
Provides support for assigning the type of scikit-learn random
forest (either `RandomForestRegressor` or `RandomForestClassifier`)
based on each feature's datatype. If the target is numeric (e.g. `int`
or `float`), a `RandomForestRegressor` is used.  If the target is
categorical (e.g. `str` or `pd.Categorical`), a `RandomForestClassifier` is
used.  Targets are automatically promoted to the minimum numpy
dtype that safely represents all elements.

- Specialize parameters for regression and classification
- Add initial auto-checking of features to set correct estimator
- Reworked mapping of column index to RF type and added initial test
- Use feature names in dictionary of RF types
- Created `utils` package for new functions
- Modify `get_feature_dtypes` to return promoted dtypes when necessary
- Add end-to-end regression tests for RFNNs with mixed forests
- Test non-default parameters for different RF forest types
- Edit `RFNNRegressor` and `RFNodeTransformer` docstrings