Implementation of weighted Hamming distance for tree-based estimators (#98)
This commit adds two new classes (`WeightedHammingDistanceMetric`
and `WeightedTreesNNRegressor`) that facilitate calculating node
dissimilarity (Hamming distance) based on weights associated with each
tree in a forest and with each forest in an estimator.

`WeightedHammingDistanceMetric` is a generic callable class that gets 
instantiated at the same time as a tree-based estimator (e.g. `RFNNRegressor`)
and is set as the function to call for the `metric` parameter for distance
finding.  It has a `set_weights` method that allows setting of the weights
after object instantiation.

`WeightedTreesNNRegressor` is an abstract base class for
`RFNNRegressor` that handles the actual setting of the Hamming
weights.  An estimator's transformer (e.g. `RFNodeTransformer`) is
responsible for setting individual tree weights on each tree in the
forests and the forest weights are a parameter on the estimator
(defaulting to "uniform"). It overrides the `_set_fitted_transformer`
method to set weights directly after the transformer has been fit.

A number of new estimator attributes were added, including
`n_forests_` and `tree_weights_` for `TreeNodeTransformer` and
`hamming_weights_` for `WeightedTreesNNRegressor`.

See #98 for more detail, but we discovered that using
`scipy.spatial.distance.hamming` led to some slight inconsistencies across
different testing platforms, likely to known imprecision in `np.dot`.