Include transformer tests for `GBNodeTransformer`
This commit also fixes one error and one warning in
`TreeNodeTransformer` found during testing.  In `transform`,
returning `X` from the call to `_validate_data` is necessary because
`GradientBoostingRegressor.apply` expects that `X` is a numpy array
(calls `.shape()`).

In `_fit`, returing `X` from the call to `_validate_data` returns
`X` as a numpy array and removes the warning about fitting with
feature names.